Changes to handle retrieval
Instead of saving a like pair to retrieve, I'll only save the DID I don't
know. I was saving the pairs because originally I thought I would register
the like event from the same place, but there's no reason to tie handle
retrieval to the registration, given we may get rate-limited on the
former but not on the later.